% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-broom.R
\name{sw_augment_columns}
\alias{sw_augment_columns}
\title{Augments data}
\usage{
sw_augment_columns(ret, data, rename_index)
}
\arguments{
\item{ret}{An object of class tibble}

\item{data}{Any time series data that is to be augmented}

\item{rename_index}{A variable indicating the index name to be used in the
tibble returned}
}
\description{
Augments data
}
